+++ src/mongo/db/pdfile.cpp
@@ -38,6 +38,7 @@ _ disallow system* manipulations from the database.
 #include "mongo/db/pdfile_private.h"
 #include "mongo/db/background.h"
 #include "mongo/db/btree.h"
+#include "mongo/db/cloner.h"
 #include "mongo/db/commands/server_status.h"
 #include "mongo/db/curop-inl.h"
 #include "mongo/db/db.h"
@@ -51,7 +52,6 @@ _ disallow system* manipulations from the database.
 #include "mongo/db/namespace-inl.h"
 #include "mongo/db/namespacestring.h"
 #include "mongo/db/ops/delete.h"
-#include "mongo/db/repl.h"
 #include "mongo/db/replutil.h"
 #include "mongo/db/sort_phase_one.h"
 #include "mongo/util/file.h"
