+++ src/mongo/db/repl/rs.cpp
@@ -24,7 +24,6 @@
 #include "mongo/db/cmdline.h"
 #include "mongo/db/dbhelpers.h"
 #include "mongo/db/instance.h"
-#include "mongo/db/repl.h"
 #include "mongo/db/repl/bgsync.h"
 #include "mongo/db/repl/connections.h"
 #include "mongo/db/repl/replication_server_status.h"  // replSettings
