+++ src/mongo/db/repl/rs_initialsync.cpp
@@ -19,9 +19,9 @@
 #include "mongo/db/repl/rs.h"
 
 #include "mongo/db/client.h"
+#include "mongo/db/cloner.h"
 #include "mongo/db/dbhelpers.h"
 #include "mongo/db/oplog.h"
-#include "mongo/db/repl.h"
 #include "mongo/db/repl/bgsync.h"
 #include "mongo/db/repl/oplogreader.h"
 #include "mongo/db/repl/optime.h"
