+++ src/mongo/db/repl/rs_rollback.cpp
@@ -22,7 +22,6 @@
 #include "mongo/db/ops/update.h"
 #include "mongo/db/ops/delete.h"
 #include "mongo/db/repl/rs.h"
-#include "mongo/db/repl.h"
 
 /* Scenarios
