+++ src/mongo/dbtests/repltests.cpp
@@ -19,14 +19,13 @@
 
 #include "mongo/pch.h"
 
-#include "mongo/db/repl.h"
-
 #include "mongo/db/db.h"
 #include "mongo/db/index_update.h"
 #include "mongo/db/instance.h"
 #include "mongo/db/json.h"
-#include "mongo/db/repl/finding_start_cursor.h"
 #include "mongo/db/queryoptimizer.h"
+#include "mongo/db/repl/finding_start_cursor.h"
+#include "mongo/db/repl/master_slave.h"
 #include "mongo/db/repl/rs.h"
 
 #include "mongo/dbtests/dbtests.h"
